+++ src/FaslReader.cpp
@@ -102,8 +102,8 @@ void FaslReader::linkShared(Object obj, EqHashTable* seen)
     }
     if (obj.isVector()) {
         Vector* v = obj.toVector();
-        int n = v->length();
-        for (int i = 0; i < n; i++) {
+        size_t n = v->length();
+        for (size_t i = 0; i < n; i++) {
             if (v->ref(i).isSharedReference()) {
                 v->set(i, getShared(v->ref(i).toSharedReference()->index()));
             } else {
